Update installer and cmake to 3.24.x  (#19268)
* Updates the engine to require CMake >= 3.24

Updates the WIX installer script and build script to include that version
of CMake downloaded from the kitware official site.

Makes the WIX post-installation steps non-failing.  They still run,
but if they fail, they do not cause the installer to rollback.

Updates the license scanner to work when licenses live in multiple
different drives on windows.  relpath in python throws an exception
if you specify paths on different drives.  This is often the case
with the 3rd party libraries (`%USERPROFILE%`) and the build
`D:/workspace` or something.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fixes extra unnecessary spaces

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>