Adds a dummy ciso646 temporarily to satisfy the really old physx packages (#19275)
* Adds a dummy ciso646 temporarily to satisfy the really old physx packages

This adds a new target called `ciso646-include` which other targets
can depend on in order to satisfy old code that looks for `<ciso646>`.

We need to phase this out, but in order to do so, we'd have to figure
out how to rebuild and patch Physx4, Physx5, and NVCloth, all of which
currently depend on this.

This gives us time to figure out what the actual plan is.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fix situations where the config is imported multiple times.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>