Cherry-pick from `stabilization/25100` to `development` (splash screen and ciso646) (#19290)
* Adds a dummy ciso646 temporarily to satisfy the really old physx packages (#19275)

This adds a new target called `ciso646-include` which other targets
can depend on in order to satisfy old code that looks for `<ciso646>`.

We need to phase this out, but in order to do so, we'd have to figure
out how to rebuild and patch Physx4, Physx5, and NVCloth, all of which
currently depend on this.

This gives us time to figure out what the actual plan is.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
* Add new splashscreen for 25.10.0 (#19286)
Signed-off-by: Mike Chang <changml@amazon.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
Signed-off-by: Mike Chang <changml@amazon.com>
Co-authored-by: Mike Chang <changml@amazon.com>