Adds additional unit tests for AZStd::copy_backward (#19284)
* Adds additional unit tests for AZStd::copy_backward

One github actions machine failed one time with
this test, in a way I can't figure out.

So I've added 4 additional tests which will clarify
what the situation is if it happens again.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Moves the check for UB out of the platform specific block.

The check for the UB needs to be right when it makes the decision
to use a memcpy / memmove operation, which is when the memory
is contiguous, and the iterators are compatible.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>